loggign graphql errors to sentry refactored middlewares code upgraded packages
