+++ package.json
@@ -75,18 +75,19 @@
     "electron": "^2.0.8",
     "electron-builder": "^20.19.2",
     "electron-builder-lib": "^20.15.3",
-    "graphql-middleware": "^1.3.0",
+    "graphql-middleware": "^2.0.2",
     "tslint": "^5.10.0",
     "typescript": "^2.8.3"
   },
   "dependencies": {
     "@sentry/electron": "^0.8.1",
+    "@types/aws-lambda": "^8.10.17",
     "args-parser": "^1.1.0",
     "auto-launch": "^5.0.5",
     "electron-log": "^2.2.15",
     "electron-store": "^2.0.0",
     "electron-updater": "^2.21.10",
-    "graphql-yoga": "^1.14.8",
+    "graphql-yoga": "^1.16.9",
     "isomorphic-git": "^0.37.3",
     "lodash": "^4.17.10",
     "parse-repo": "^1.0.4",
